Add README.md file with a basic home-manager based NixOS configuration
A new README.md file has been added with a basic home-manager based NixOS configuration. This configuration was created during the user's exploration into the NixOS world and is intended for personal use only. It is not recommended to use this configuration as an inspiration. Please note that this configuration is very basic, not safe, and is not actively maintained. The user has stated that they will not continue using or developing this configuration.

Changes:
- Added README.md with a basic home-manager based NixOS configuration.