Update NixOS personal configuration and README.md
This commit updates the NixOS personal configuration and README.md file. The changes include adding instructions on how to use the configuration, specifically cloning the repository, running NixOS-rebuild and Home Manager, and a disclaimer stating that the configuration is not actively maintained. These changes aim to improve the clarity and usability of the configuration for potential users.